fix: Identity wizard input state bug.
